\name{tsetcovf}
\alias{tsetcovf}
\title{Temporal set covariance functions}
\description{The temporal set covariance functions for a interval time T have a general approximation formula.}

\references{
Rodriguez-Cortes, F. J., Ghorbani, M., Mateu, J. and Stoyan, D. (2014). On the expected value and variance for an estimator of the spatio-temporal product density function. Department of Mathematical Sciences, Aalborg University. 19 p. (Research Report Series; No. R-2014-03).
}